pyproject.toml
[project]
name = "crypto-agent-forecaster"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai>=0.121.1",
    "langchain>=0.3.25",
    "langchain-anthropic>=0.3.14",
    "langchain-google-genai>=2.1.5",
    "langchain-openai>=0.3.18",
    "numpy>=2.2.6",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "rich>=13.9.4",
    "ta>=0.11.0",
    "typer>=0.16.0",
]
